requirements.txt
anyio==3.5.0
asgiref==3.5.0
backports.zoneinfo==0.2.1
boto3==1.20.46
botocore==1.23.46
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
defusedxml==0.7.1
Django==3.2.10
django-allauth==0.47.0
django-cors-headers==3.11.0
django-filter==2.4.0
django-heroku==0.3.1
django-mysql==4.5.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
dnspython==2.2.0
gunicorn==20.0.4
idna==3.3
importlib-metadata==4.10.1
jmespath==0.10.0
Markdown==3.3.6
mysql-connector-python==8.0.28
oauthlib==3.1.1
Pillow==9.0.0
protobuf==3.19.4
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
pymongo==4.0.1
PyMySQL==1.0.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
s3transfer==0.5.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.18.0
typing_extensions==4.0.1
urllib3==1.26.8
whitenoise==6.0.0
zipp==3.7.0
